pyproject.toml
[build-system]
requires      = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["./src"]  # list of folders that contain the packages (["."] by default)
include = ["*"]  # package names should match these glob patterns (["*"] by default)

[tool.setuptools.package-data]
"*" = ["*.yaml"]

[project]
name = "generate-gizmosql-token"
version = "0.0.14"
description = "A package which generates a GizmoSQL token (JWT) for auth testing"
readme = "README.md"
authors = [{ name = "Philip Moore", email = "philip@gizmodata.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["gizmosql", "auth", "token", "bearer", "jwt"]
dependencies = [
    "click==8.2.*",
    "pyjwt==2.10.*",
    "cryptography==45.0.*",
    "python-dotenv==1.1.*",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "bumpver",
    "pip-tools",
    "pytest>=8.0",
]
integration = [
    "docker",
    "adbc-driver-gizmosql",
    "pyarrow",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests that require Docker and a running GizmoSQL container (deselect with '-m \"not integration\"')",
]

[project.urls]
Homepage = "https://github.com/gizmodata/generate-gizmosql-token"

[project.scripts]
generate-gizmosql-token = "generate_gizmosql_token.main:click_generate_gizmosql_token"

[tool.bumpver]
current_version = "0.0.14"
version_pattern = "MAJOR.MINOR.PATCH[PYTAGNUM]"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    '^version = "{version}"$',
    '^current_version = "{version}"$',
]
"src/generate_gizmosql_token/__init__.py" = [
    '^__version__ = "{version}"$',
]
